requirements.txt
anyio==3.6.2
async-asgi-testclient==1.4.11
async-generator==1.10
attrs==23.1.0
black==23.3.0
bson==0.5.10
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pyasn1==0.4.8
pydantic==1.10.7
pydantic-mongo==1.0.1
PyJWT==2.6.0
pymongo==4.3.3
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.28.2
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.26.1
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
webdriver-manager==3.8.6
wsproto==1.2.0
